VPN Beginner’s Guide by John Mason added
I added the John Mason's Guide because it answers the most asked questions and it's fantastic written. The guide doesn't tell you which VPN Provider you should buy/support, moreover (and I think it's the better way) everyone can read it and understand what are the points for choosing a 'good' provider.